virtualized cell_group
