+++ .ycm_extra_conf.py
@@ -36,7 +36,7 @@ import ycm_core
 # CHANGE THIS LIST OF FLAGS. YES, THIS IS THE DROID YOU HAVE BEEN LOOKING FOR.
 flags = [
     '-DNDEBUG',
-    '-DNMC_HAVE_TBB',
+    '-DNMC_HAVE_CTHREAD',
     '-std=c++11',
     '-x',
     'c++',
