+++ src/sampler_function.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <functional>
+
+#include <common_types.hpp>
+#include <util/optional.hpp>
+
+namespace nest {
+namespace mc {
+
+using sampler_function = std::function<util::optional<time_type>(time_type, float)>;
+
+} // namespace mc
+} // namespace nest
