+++ tests/validation/validate_ball_and_stick.cu
@@ -3,24 +3,24 @@
 
 #include <fvm_multicell.hpp>
 
-using lowered_cell = nest::mc::fvm::fvm_multicell<nest::mc::gpu::backend>;
+const auto backend = nest::mc::backend_policy::prefer_gpu;
 
 TEST(ball_and_stick, neuron_ref) {
-    validate_ball_and_stick<lowered_cell>();
+    validate_ball_and_stick(backend);
 }
 
 TEST(ball_and_taper, neuron_ref) {
-    validate_ball_and_taper<lowered_cell>();
+    validate_ball_and_taper(backend);
 }
 
 TEST(ball_and_3stick, neuron_ref) {
-    validate_ball_and_3stick<lowered_cell>();
+    validate_ball_and_3stick(backend);
 }
 
 TEST(rallpack1, numeric_ref) {
-    validate_rallpack1<lowered_cell>();
+    validate_rallpack1(backend);
 }
 
 TEST(ball_and_squiggle, neuron_ref) {
-    validate_ball_and_squiggle<lowered_cell>();
+    validate_ball_and_squiggle(backend);
 }
