+++ tests/validation/validate_soma.cu
@@ -2,8 +2,8 @@
 
 #include "../gtest.h"
 
-using lowered_cell = nest::mc::fvm::fvm_multicell<nest::mc::gpu::backend>;
+const auto backend = nest::mc::backend_policy::prefer_gpu;
 
 TEST(soma, numeric_ref) {
-    validate_soma<lowered_cell>();
+    validate_soma(backend);
 }
