+++ tests/validation/validate_synapses.cu
@@ -3,14 +3,14 @@
 #include "../gtest.h"
 #include "validate_synapses.hpp"
 
-using lowered_cell = nest::mc::fvm::fvm_multicell<nest::mc::gpu::backend>;
+const auto backend = nest::mc::backend_policy::prefer_gpu;
 
 TEST(simple_synapse, expsyn_neuron_ref) {
     SCOPED_TRACE("expsyn");
-    run_synapse_test<lowered_cell>("expsyn", "neuron_simple_exp_synapse.json");
+    run_synapse_test("expsyn", "neuron_simple_exp_synapse.json", backend);
 }
 
 TEST(simple_synapse, exp2syn_neuron_ref) {
     SCOPED_TRACE("exp2syn");
-    run_synapse_test<lowered_cell>("exp2syn", "neuron_simple_exp2_synapse.json");
+    run_synapse_test("exp2syn", "neuron_simple_exp2_synapse.json", backend);
 }
